0.999
This is the most stable and complete version to date, I played around with the friendship bracelet mechanic. ALSO completed a lot of the early roomdescriptions, and even wrote in a talking cat ;) PLEASE PLEASE let me know if you need to understand anything, also remember to use Display Commands and display hidden commands. Happy hunting